rename the `childCloser` interface
Rename the `childCloser` interface to `mutableParent` to better reflect the fact
that the structures that implement it (`Directory` and `Root`) have the function
of parents in the MFS hierarchy. Rename also the method of the interface from
`closeChild` to `updateChildEntry` shifting the focus away from in which
circumstance is the method being called (when closing a child) to actually what
operation is being performed by the method (updating an entry in the parent to
the new content).