remove `Sync` from `FileDescriptor`
Remove the `Sync()` method from the `FileDescriptor` and also its
implementation.

Two different undocumented methods in `FileDescriptor` are offered to update the
state of the MFS: `Flush()` and `Sync()`, both calling `updateChildEntry`. The
only difference (the user is not aware of) is that one sets the `fullSync`
argument and the other doesn't, that is, one does a full update of the
filesystem and the other just updates the parent directory.

This current situation is not clear to the consumer who may use one desiring the
effect of the other. As a precautionary measure while redesigning the MFS API
one method is removed to simplify its usage. The `Sync()` was chosen as the less
"safe" alternative which didn't do a full update of the MFS and also doesn't
seem to be used in the `go-ipfs` repository (the main consumer of the `mfs`
package).