remove the `fullSync` option from `updateChildEntry`
Make `updateChildEntry` always propagate the update all the way up to the root,
the equivalent of calling it always with `fullSync` set.

The case of calling it without setting `fullSync` (a kind of "half-update")
where only the parent's directory UnixFS node was updated (but nothing else,
leaving the root outdated) seemed of little used. This helps to simplify the
logic around the update mechanism in MFS.