provide a `ListEntries` method
The `coreunix.Adder` was calling the `ListNames` method and then converting the
list to `FSNode`s with `Child`. The `Child` method generates a cache entry for
each child retrieved, to avoid reduce memory use the `Adder` called the exported
`Uncache` method (added specifically for that case) which gave outside access to
the internal cache. For more details see: https://github.com/ipfs/go-
ipfs/pull/3027.

To prevent access to the internal cache a `ListEntries` method is provided with
the same logic extracted from the `coreunix.Adder`. This allows now to make the
`Uncache` method private, which will permit a cache refactor in the future
without breaking the API.