fix/32/pr-ports : Pulled changes from PR 4517
Pulled changes from https://github.com/ipfs/go-ipfs/pull/4517, on top of, https://github.com/ipfs/go-mfs/pull/45.
Change added to unblock the `waitPub()` call. With the elimination of
stateSync cause a `updateChildEntry` to happen for `stateFlushed` as
well, causing it to propogate upwards to the parent(s) [fullSync] and
force a publish to happen, hence unblocking `waitPub`.