Do not store context in structs
Stop storing context in structs, and instead pass a context as a function argument in the places where support for cancellation is needed.

Much of the functionality does not need to support concelation, and even in some cases where a request may have been canceled, letting the function complete is preferable to avoid parital updates to MFS.  Cancellation is only needed in places where a caller may be waiting for a signal that something has completed.

- Handle additional TODO items in code
- Fix potential race condition
- Close() always stops republisher; fixes possible goroutine leak
- Republisher can pick up new values during retry loop.